+++ src/store/myApps/state.js
@@ -1,0 +1,9 @@
+export default {
+  configuredApps: [],
+  loadingConfiguredApps: false,
+  errorConfiguredApps: null,
+
+  installedApps: [],
+  loadingInstalledApps: false,
+  errorInstalledApps: null,
+};